% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as-helpers}
\alias{as-helpers}
\alias{as_ugm3}
\alias{as_person}
\alias{as_population}
\alias{as_mortality}
\alias{as_mortality_rate}
\alias{as_CANCRISK}
\title{Set commonly-used units}
\usage{
as_ugm3(x)

as_person(x)

as_population(x)

as_mortality(x)

as_mortality_rate(x, u = "death/Mperson/yr")

as_CANCRISK(x, u = "cancer/Mperson")
}
\description{
Set commonly-used units
}
