% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_fc.R
\name{curve_fc}
\alias{curve_fc}
\title{curve_fc}
\usage{
curve_fc(model, par, col = "red2", add = FALSE, leg = TRUE)
}
\arguments{
\item{model}{character. The model we want to curve.}

\item{par}{numeric. A vector of parameters.}

\item{col}{character. The curve color.}

\item{add}{logic. Should we allow more curves over?}

\item{leg}{logic. Show we show the legends?}
}
\value{
A curve.
}
\description{
A function to make curves of the distributions.
}
\examples{
{
curve_fc("beta", par = c(0.5, 0.5))
}
}
