% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBeta.R
\name{dBeta}
\alias{dBeta}
\title{dBeta}
\usage{
dBeta(y, mu, phi)
}
\arguments{
\item{y}{numeric. Variable of interest.}

\item{mu}{numeric. Beta's first parameter.}

\item{phi}{numeric. Beta's second parameter.}
}
\value{
Density values.
}
\description{
Beta distribution.
}
\examples{
{
\donttest{
  curve(dBeta(y = x, mu = 0.5, phi = 0.5))
}
}
}
