% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBR.R
\name{dBR}
\alias{dBR}
\title{dBR}
\usage{
dBR(y, gamma, phi, alpha)
}
\arguments{
\item{y}{numeric. Variable of interest.}

\item{gamma}{numeric. RB's first parameter.}

\item{phi}{numeric. RB's second parameter.}

\item{alpha}{numeric. RB's third parameter.}
}
\value{
Density values.
}
\description{
Retangular Beta distribution.
}
\examples{
{
\donttest{
  curve(dBR(y = x, gamma = 0.5, phi = 0.5, alpha = 0.5))
}
}
}
