% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgUnit.R
\name{dgUnit}
\alias{dgUnit}
\title{dgUnit}
\usage{
dgUnit(y, mu, phi)
}
\arguments{
\item{y}{numeric. Variable of interest.}

\item{mu}{numeric. Unit Gamma's first parameter.}

\item{phi}{numeric. Unit Gamma's second parameter.}
}
\value{
Density values.
}
\description{
Unit Gamma distribution.
}
\examples{
{
\donttest{
  curve(dgUnit(y = x, mu = 0.5, phi = 0.5))
}
}
}
