% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dKum.R
\name{dKum}
\alias{dKum}
\title{dKum}
\usage{
dKum(y, mu, phi, a = 0, b = 1)
}
\arguments{
\item{y}{numeric. Variable of interest.}

\item{mu}{numeric. Kumaraswamy's first parameter.}

\item{phi}{numeric. Kumaraswamy's second parameter.}

\item{a}{numeric.}

\item{b}{numeric.}
}
\value{
Density values.
}
\description{
Kumaraswamy distribution.
}
\examples{
{
\donttest{
  curve(dKum(y = x, mu = 0.5, phi = 0.5))
}
}
}
