% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dLEP.R
\name{dLEP}
\alias{dLEP}
\title{dLEP}
\usage{
dLEP(y, mu, phi, nu)
}
\arguments{
\item{y}{numeric. Variable of interest.}

\item{mu}{numeric. LEP's first parameter.}

\item{phi}{numeric. LEP's second parameter.}

\item{nu}{numeric. LEP's third parameter.}
}
\value{
Density values.
}
\description{
Logit Exponential Power distribution.
}
\examples{
{
\donttest{
  curve(dLEP(y = x, mu = 0, phi = 0.5, nu = 0.5))
}
}\\
}
