% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dLL.R
\name{dLL}
\alias{dLL}
\title{dLL}
\usage{
dLL(y, mu, phi)
}
\arguments{
\item{y}{numeric. Variable of interest.}

\item{mu}{numeric. LL's first parameter.}

\item{phi}{numeric. LL's second parameter.}
}
\value{
Density values.
}
\description{
Logit Logistic distribution.
}
\examples{
{
\donttest{
  curve(dLL(y = x, mu = 0, phi = 0.5))
}
}
}
