% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dLN.R
\name{dLN}
\alias{dLN}
\title{dLN}
\usage{
dLN(y, mu, phi)
}
\arguments{
\item{y}{numeric. Variable of interest.}

\item{mu}{numeric. LN's first parameter.}

\item{phi}{numeric. LN's second parameter.}
}
\value{
Density values.
}
\description{
Logit Normal distribution.
}
\examples{
{
\donttest{
  curve(dLN(y = x, mu = 0.5, phi = 0.5))
}
}
}
