% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dLSEP.R
\name{dLSEP}
\alias{dLSEP}
\title{dLSEP}
\usage{
dLSEP(y, mu, phi, lambda, tau)
}
\arguments{
\item{y}{numeric. Variable of interest.}

\item{mu}{numeric. LSEP's first parameter.}

\item{phi}{numeric. LSEP's second parameter.}

\item{lambda}{numeric. LSEP's third parameter.}

\item{tau}{numeric. LSEP's fourth parameter.}
}
\value{
Density values.
}
\description{
Logit Skewed Exponential Power distribution.
}
\examples{
{
\donttest{
  curve(dLSEP(y = x, mu = 0, phi = 0.5, lambda = 0.5, tau = 0.5))
}
}
}
