% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dLSL.R
\name{dLSL}
\alias{dLSL}
\title{dLSL}
\usage{
dLSL(y, mu, phi, lambda)
}
\arguments{
\item{y}{numeric. Variable of interest.}

\item{mu}{numeric. LSL's first parameter.}

\item{phi}{numeric. LSL's second parameter.}

\item{lambda}{numeric. LSL's third parameter.}
}
\value{
Density values.
}
\description{
Logit Skewed Logistic distribution.
}
\examples{
{
\donttest{
  curve(dLSL(y = x, mu = 0.5, phi = 0.5, lambda = 0.5))
}
}
}
