% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dLSN.R
\name{dLSN}
\alias{dLSN}
\title{dLSN}
\usage{
dLSN(y, mu, phi, lambda)
}
\arguments{
\item{y}{numeric. Variable of interest.}

\item{mu}{numeric. LSN's first parameter.}

\item{phi}{numeric. LSN's second parameter.}

\item{lambda}{numeric. LSN's third parameter.}
}
\value{
Density values.
}
\description{
Logit Skewed Normal distribution.
}
\examples{
{
\donttest{
  curve(dLSN(y = x, mu = 0, phi = 0.5, lambda = 0.5))
}
}
}
