% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dLST.R
\name{dLST}
\alias{dLST}
\title{dLST}
\usage{
dLST(y, mu, phi, lambda, tau)
}
\arguments{
\item{y}{numeric. Variable of interest.}

\item{mu}{numeric. LST's first parameter.}

\item{phi}{numeric. LST's second parameter.}

\item{lambda}{numeric. LST's third parameter.}

\item{tau}{numeric. LST's fourth parameter.}
}
\value{
Density values.
}
\description{
Logit Skewed t distribution.
}
\examples{
{
\donttest{
  curve(dLST(y = x, mu = 0, phi = 0.5, lambda = 0.5, tau = 0.5))
}
}
}
