% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dLT.R
\name{dLT}
\alias{dLT}
\title{dLT}
\usage{
dLT(y, mu, phi, nu)
}
\arguments{
\item{y}{numeric. Variable of interest.}

\item{mu}{numeric. LST's first parameter.}

\item{phi}{numeric. LST's second parameter.}

\item{nu}{numeric. LST's third parameter.}
}
\value{
Density values.
}
\description{
Logit t distribution.
}
\examples{
{
\donttest{
  curve(dLT(y = x, mu = 0, phi = 0.5, nu = 0.5))
}
}
}
