% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimplex.R
\name{dSimplex}
\alias{dSimplex}
\title{dSimplex}
\usage{
dSimplex(y, mu, phi)
}
\arguments{
\item{y}{numeric. Variable of interest.}

\item{mu}{numeric. Simplex's first parameter.}

\item{phi}{numeric. Simplex's second parameter.}
}
\value{
Density values.
}
\description{
Simplex distribution.
}
\examples{
{
\donttest{
  curve(dSimplex(y = x, mu = 0.5, phi = 0.5))
}
}
}
