% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mortality_rate.R
\name{format_mortality_rate}
\alias{format_mortality_rate}
\title{format_mortality_rate}
\usage{
format_mortality_rate(
  x,
  units = "death/Mperson",
  digits = NULL,
  signif = Inf,
  suffix = NULL
)
}
\arguments{
\item{x}{object that \emph{might} be a \code{units} object}

\item{units}{(optional) character, like "death/Mperson"}

\item{digits}{(optional) integer}

\item{signif}{(optional) integer number of digits, as in \code{\link[base:Round]{base::signif()}}}

\item{suffix}{(optional) character}
}
\value{

}
\description{
format_mortality_rate
}
