% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_fit.R
\name{g_fit}
\alias{g_fit}
\title{g_fit}
\usage{
g_fit(model, y)
}
\arguments{
\item{model}{character. The chosen model.}

\item{y}{numeric. The variable values.}
}
\value{
the fitted parameters
}
\description{
Fits `gamlss` distributions. The options are "ll", "ln", "lt",
"lep", "lsn", "lst", "lsep"
}
\examples{
{
\donttest{
  x <- seq(0.00001, 0.9999, l = 50)
  g_fit("ll", x)
}
}
}
