% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_units.R
\name{get_units}
\alias{get_units}
\title{get_units}
\usage{
get_units(x)
}
\description{
Works like \code{\link[units:deparse_unit]{units::deparse_unit()}}, except it returns \code{NULL} by
default, rather than throwing an error.
}
\details{
@param x object
}
\seealso{
\code{\link[=has_units]{has_units()}}
}
