% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_units.R
\name{has_units}
\alias{has_units}
\title{has_units}
\usage{
has_units(x, u = NULL)
}
\arguments{
\item{x}{(vector) to be tested}

\item{u}{(optional) expected units}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Returns \code{TRUE} if \code{x} is either (a) unitless, or is not of class "units".
If \code{x} is of class "units", can test whether the units are equal to \code{u}.
}
