% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike.R
\name{loglike}
\alias{loglike}
\title{loglike}
\usage{
loglike(par, model, x, sum = "yes")
}
\arguments{
\item{par}{numeric. Parameters for the distribution.}

\item{model}{character. The chosen model.}

\item{x}{numeric. The variable values.}

\item{sum}{logic. Should we return the sum of the 
loglikelihood?}
}
\value{
loglikelihood values
}
\description{
Loglikelihood for the Beta, Kumaraswamy, Unit Gamma, Simplex,
 Retangular Beta or LSL distributions.
}
\examples{
{
\donttest{
  x <- seq(0.00001, 0.9999, l = 50)
  loglike(par = c(0.5, 0.5), model = "beta", x = x, sum = "yes")
}
}
}
