% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{quality}
\alias{quality}
\title{quality}
\usage{
quality(ll, n_par)
}
\arguments{
\item{ll}{numeric. The loglikehood results.}

\item{n_par}{numeric. The number of parameters.}
}
\value{
AIC and BIC values
}
\description{
Calculates the AIC and BIC for the chosen model.
}
\examples{
{
\donttest{
  x <- seq(0.00001, 0.9999, l = 50)
  ll <- loglike(par = c(0.5, 0.5), model = "beta", x = x, sum = "yes")
  quality(ll, n_par = 2)
}
}
}
