% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_units.R
\name{restore_units}
\alias{restore_units}
\title{Restore units to tabular data}
\usage{
restore_units(to, from)
}
\arguments{
\item{to}{tabular data or geodata}

\item{from}{tabular data or geodata}
}
\value{
same as \code{from}, but with units set on columns as they were in \code{to}
}
\description{
Operations on tabular data or geodata can result in unit-aware columns being returned to a naive state.
This function easily restores them.
}
\examples{
\dontrun{
library(dplyr)
library(units)
df <- mutate(mtcars, mpg = as_units(mpg, "mi/gal"))
grouped <- group_by(df, cyl)
summarised <- summarise_at(grouped, vars(mpg), mean)
restore_units(summarised, from = df)
}

}
