% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_install_unit.R
\name{try_install_unit}
\alias{try_install_unit}
\title{try_install_unit}
\usage{
try_install_unit(...)
}
\arguments{
\item{...}{passed to \code{\link[units:install_unit]{units::install_unit()}}}
}
\description{
Just like \code{units::install_unit()}, but re-raises error(s) as warning(s).
}
\note{
The \code{units} package used to support \code{install_symbolic_unit(..., warn = TRUE)}.
With that syntax, attempting to install an already-installed unit just led to a warning;
it didn't throw an error. This is a way of recreating that behavior.
There doesn't seem to be a function offered by \code{units} to check whether a symbolic
unit is already installed. If there were, that would seem to be more prefereable than
a blanket \code{tryCatch()}.
}
