% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mortality_rate.R
\name{format_mortality_rate}
\alias{format_mortality_rate}
\title{format_mortality_rate}
\usage{
format_mortality_rate(
  x,
  units = "death/Mperson",
  digits = NULL,
  signif = Inf,
  suffix = NULL
)
}
\arguments{
\item{suffix}{}
}
\value{

}
\description{
format_mortality_rate
}
